Added filtration component

+++ server/models/models.js
@@ -13,7 +13,7 @@ const User = sequelize.define('user', {
 const Item = sequelize.define('item', {
     id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
     name: { type: DataTypes.STRING, allowNull: false },
-    price: { type: DataTypes.INTEGER, allowNull: false },
+    price: { type: DataTypes.DOUBLE, allowNull: false },
     typeId: {type: DataTypes.INTEGER, allowNull: false},
     rating: { type: DataTypes.INTEGER, defaultValue: 0 },
     img: { type: DataTypes.STRING, allowNull: false },
